🧹 chore: organize code
