--- sources/@roots/bud-api/tsconfig.test.json
@@ -1,18 +1,0 @@
-{
-  "extends": "../../../config/tsconfig.test.json",
-  "compilerOptions": {
-    "noEmit": true,
-    "strict": true,
-    "rootDir": "./src",
-    "outDir": "./lib"
-  },
-  "include": ["*.test.ts"],
-  "exclude": ["./lib", "./node_modules"],
-  "references": [
-    {"path": "../bud-build/tsconfig.json"},
-    {"path": "../bud-extensions/tsconfig.json"},
-    {"path": "../bud-framework/tsconfig.json"},
-    {"path": "../bud-support/tsconfig.json"},
-    {"path": "../bud-terser/tsconfig.json"}
-  ]
-}
